requirements.txt
beaker==1.11.0
pyramid==1.10.4
pyramid_jinja2==2.8
pyramid_debugtoolbar==4.6.1
pyramid_beaker==0.8
pyramid_tm==2.4
colander==1.7.0
deform==2.0.10
Babel==2.8.0
lingua==4.14
pymongo==3.11.0
py-bcrypt==0.4
# Gunincorn 20 is not compatible with python 2.7
gunicorn==19.10.0
celery==4.4.7
pyramid_celery==3.0.0
cornice==4.0.1
jsonschema==3.2.0
gevent==20.6.2
gevent-websocket==0.10.1
python-socketio==4.6.0
redis==3.5.3
WebHelpers2==2.0
paginate==0.5.6
Paste==3.4.3
PyChef==0.3.0
xmltodict==0.12.0
requests==2.24.0
beautifulsoup4==4.9.1
iscompatible==0.1.1
pycrypto
xhtml2pdf
python-dateutil==2.8.1
# Zipp >=2.0.0 is not compatible with for python 2.7
zipp==1.2.0 
# Configparser >= 5.0.0 is not compatible with for python 2.7
configparser==4.0.2
# Pillow>=7 is not compatible with for python 2.7
Pillow==6.2.2
# Venusian >=3.0.0 is not compatible with for python 2.7
venusian==2.1.0
# Force versions to not to use alpha pre-releases
MarkupSafe==1.1.1 
Jinja2==2.11.2 
